+++ src/common.rs
@@ -350,6 +350,7 @@ impl Observed {
 }
 
 #[inline]
+#[optimize(speed)]
 unsafe fn _get_state(py: pyo3::Python<'_>, ptr: *mut pyo3::ffi::PyObject) -> pyo3::PyResult<u16> {
     unsafe fn inner(
         py: pyo3::Python<'_>,
@@ -485,6 +486,4 @@ impl TimeToLivePair {
     }
 }
 
-// Thanks to `Amanieu d'Antras` for this beautiful implementation.
-// https://github.com/Amanieu/parking_lot/blob/eeb186c48c8e6433c10f7552ef1cd1d56e5c83b1/src/raw_mutex.rs
-pub type Mutex<T> = lock_api::Mutex<parking_lot::RawMutex, T>;
+pub type Mutex<T> = parking_lot::Mutex<T>;