+++ src/linked_list.rs
@@ -78,6 +78,7 @@ impl LinkedList {
         }
     }
 
+    #[inline]
     pub fn clear(&mut self) {
         while self.pop_front().is_some() {}
     }
@@ -144,6 +145,7 @@ impl LinkedList {
         self.tail = Some(node);
     }
 
+    #[inline]
     pub fn iter(&self) -> Iter {
         Iter {
             head: self.head,