+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v5.0.3 - 2025-10-10
+### Changed
+- Better support for Python 3.14
+- Use `immutable_type` feature for Python 3.14
+- Some improvements
+
+### Thanks
+- thanks to [@chirizxc](https://github.com/chirizxc)
+
+### Internal
+- Bump rust dependencies
+- Use Rust nightly
+- Use `parking_lot::Mutex` instead of duplicating its code
+
 ## v5.0.2 - 2025-09-14
 ### Changed
 - Support Python 3.14 ( for linux )